move tests to module
